refactor(LibraryManager): match namespace to "Helpers" folder.
