+++ src/LibraryManager/Helpers/CancellationHelpers.cs
@@ -4,7 +4,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.Web.LibraryManager
+namespace Microsoft.Web.LibraryManager.Helpers
 {
     internal static class CancellationHelpers
     {
