+++ src/LibraryManager/Providers/Unpkg/NpmPackageSearch.cs
@@ -5,6 +5,7 @@ using System.Globalization;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Web;
+using Microsoft.Web.LibraryManager.Helpers;
 using Newtonsoft.Json.Linq;
 
 #if NET472
