+++ test/LibraryManager.Test/ExtensionsTest.cs
@@ -4,6 +4,7 @@
 using System.Collections.Generic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.Web.LibraryManager.Contracts;
+using Microsoft.Web.LibraryManager.Helpers;
 using Microsoft.Web.LibraryManager.Mocks;
 
 namespace Microsoft.Web.LibraryManager.Test
